shop_최준서 feat: 마이페이지 리뷰하기, 리뷰 수정 로직 추가

+++ src/main/java/com/app/bluecotton/service/ShopServiceImpl.java
@@ -8,12 +8,11 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Map;
-
 @Service
 @Transactional(rollbackFor = Exception.class)
 @RequiredArgsConstructor
-
 public class ShopServiceImpl implements ShopService {
+
 
     private final ShopDAO shopDAO;
 
@@ -23,8 +22,8 @@ public class ShopServiceImpl implements ShopService {
     }
 
     @Override
-    public ProductDetailResponseDTO getProductDetailHeader(Long id) {
-        return shopDAO.findProductDetailHeader(id);
+    public ProductDetailResponseDTO getProductDetailHeader(Long id, Long memberId) {
+        return shopDAO.findProductDetailHeader(id, memberId);
     }
 
     @Override
@@ -33,7 +32,7 @@ public class ShopServiceImpl implements ShopService {
     }
 
     @Override
-    public List<ProductReviewDetailResponseDTO> getProductReviewDetail(Map<String,Object> reviewParams) {
+    public List<ProductReviewDetailResponseDTO> getProductReviewDetail(Map<String, Object> reviewParams) {
         return shopDAO.findProductReviewDetail(reviewParams);
     }
 
@@ -42,14 +41,27 @@ public class ShopServiceImpl implements ShopService {
         return shopDAO.findProductReviewStats(id);
     }
 
+    // 찜하기 토글
     @Override
-    public List<ProductListResponseDTO> getLikedProducts(Long memberId) {
-        return shopDAO.findLikedProducts(memberId);
+    @Transactional
+    public void toggleLike(Long memberId, Long productId) {
+        Integer count = shopDAO.findLikeCount(memberId, productId);
+
+        // 찜한 상품이 있을 때
+        if (count > 0) {
+            // 찜 삭제
+            shopDAO.deleteLikedProduct(memberId, productId);
+        }
+        // 찜한 상품 없을 때
+        else {
+            // 찜 추가
+            shopDAO.insertMyLikedProduct(memberId, productId);
+        }
     }
 
     @Override
-    public void unLikeProduct(Long memberId, Long productId) {
-        shopDAO.deleteLikedProduct(memberId, productId);
+    public List<ProductListResponseDTO> getLikedProducts(Long memberId) {
+        return shopDAO.findLikedProducts(memberId);
     }
 
     @Override
@@ -57,5 +69,39 @@ public class ShopServiceImpl implements ShopService {
         return shopDAO.findMyReviews(id);
     }
 
+    @Override
+    public void modifyMyReview(Map<String, Object> modifyReview) {
+        shopDAO.updateMyReview(modifyReview);
+    }
+
+    @Override
+    public void deleteMyReview(Long id) {
+        shopDAO.deleteMyReview(id);
+    }
+
+    @Override
+    public List<MyPageOrderListDTO> getMyOrders(Long memberId) {
+        return shopDAO.findMyOrders(memberId);
+    }
+
+    @Override
+    public Map<String, Object> getReviewModal(Long productId) {
+        return shopDAO.getReviewModal(productId);
+    }
+
+    @Override
+    public void insertMyReview(MyPageReviewWriteDTO myPageReviewWriteDTO) {
+        shopDAO.insertMyReview(myPageReviewWriteDTO);
+    }
+
+    @Override
+    public void insertMyReviewImage(MyPageReviewWriteDTO myPageReviewWriteDTO) {
+        shopDAO.insertMyReviewImage(myPageReviewWriteDTO);
+    }
+
 
 }
+
+
+
+
